chore: rename section index components to 'index'
