+++ src/pages/onboarding/index.tsx
@@ -1,12 +1,12 @@
 import Navbar from "src/components/common/navbar";
 
-import LapTopOpening from "./sections/1_LapTopOpening/LapTopOpening";
-import Introduction from "./sections/2_Introduction/Introduction";
-import CoolDevCulture from "./sections/3_CoolDevCulture/CoolDevCulture";
+import LapTopOpening from "./sections/1_LapTopOpening";
+import Introduction from "./sections/2_Introduction";
+import CoolDevCulture from "./sections/3_CoolDevCulture";
 
 const OnboardingPage = () => {
   return (
-    <>
+    <div className="w-full overflow-x-hidden">
       <div className="fixed top-0 left-0 w-full z-10 content-wrapper">
         <Navbar mode="dark" hasBackground={false} />
       </div>
@@ -19,7 +19,7 @@ const OnboardingPage = () => {
 
         <CoolDevCulture />
       </main>
-    </>
+    </div>
   );
 };
