+++ src/pages/onboarding/sections/1_LapTopOpening/index.tsx
