+++ src/pages/onboarding/sections/2_Introduction/index.tsx
