+++ src/pages/onboarding/sections/3_CoolDevCulture/index.tsx
